Lane Detection 1st Demo program
Hi All,
This program basically targets on highlighting the White Lines on the road. This can be termed as LANE detection as the lanes are separated via the White Lines between them.
Quick brain teaser: What if RGB == 0, all will be black; if RGB == 255, all can be white?
Test out the program and see the output yourself.